feat(api-domain): refactor api and domain connection
