+++ src/main/java/com/jarrvis/ticketbooking/domain/Ticket.java
@@ -1,11 +1,14 @@
 package com.jarrvis.ticketbooking.domain;
 
 import lombok.AllArgsConstructor;
+import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-@Getter
+@Data
 @AllArgsConstructor
+@NoArgsConstructor
 @EqualsAndHashCode
 public class Ticket {
