+++ src/main/java/com/jarrvis/ticketbooking/ui/dto/response/ReservationResource.java
@@ -2,6 +2,7 @@ package com.jarrvis.ticketbooking.ui.dto.response;
 
 import com.jarrvis.ticketbooking.domain.ReservationStatus;
 import com.jarrvis.ticketbooking.domain.Ticket;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -9,9 +10,11 @@ import java.time.LocalDateTime;
 import java.util.Set;
 
 @Data
+@AllArgsConstructor
 @NoArgsConstructor
 public class ReservationResource {
 
+    private String reservationId;
     private String token;
     private ReservationStatus status;
     private LocalDateTime expiresAt;
