+++ src/main/java/com/jarrvis/ticketbooking/ui/dto/response/RoomResource.java
@@ -1,5 +1,6 @@
 package com.jarrvis.ticketbooking.ui.dto.response;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -7,6 +8,7 @@ import java.time.LocalDateTime;
 
 @Data
 @NoArgsConstructor
+@AllArgsConstructor
 public class RoomResource {
 
     private String name;
