+++ src/main/java/com/jarrvis/ticketbooking/ui/dto/response/ScreeningResource.java
@@ -1,5 +1,6 @@
 package com.jarrvis.ticketbooking.ui.dto.response;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -8,14 +9,15 @@ import java.util.List;
 import java.util.Map;
 
 @Data
+@AllArgsConstructor
 @NoArgsConstructor
 public class ScreeningResource {
 
     private String id;
     private String movie;
+    private String room;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
-    private String room;
     private int rows;
     private int seatsPerRow;
     private Map<Integer, List<Integer>> freePlaces;
